Import prescription data into MongoDB
Our ultimate goal is to get the data into a [Kafka][1] data pipeline,
on [Confluent Cloud][2]. Confluent Cloud does not provide an out of the
box way of consuming JSON data (they instead require you to set up
additional infrastructure yourself). To avoid that overhead and to keep
things as simple as we can, we are using [MongoDB's Atlas cloud][3]
as our go-between for the JSON data and Kakfa. MongoDB is a good choice
of go-between as it:

1. [has an import tool which makes importing JSON easy][4]
2. [is a first-class connector on the Confluent Cloud platform][5],
which makes it straightforward to get the data from MongoDB into
Kafka.

[1]: https://kafka.apache.org/
[2]: https://www.confluent.io/confluent-cloud
[3]: https://www.mongodb.com/cloud/atlas
[4]: https://docs.mongodb.com/guides/server/import/
[5]: https://www.confluent.io/blog/confluent-cloud-offers-kafka-mongodb-connector-fully-managed-clusters/