Import prescription data into MongoDB (#11)
* Install MongoDB shell on remote container

[Mongo shell documentation][1]

[1]: https://docs.mongodb.com/manual/reference/program/mongo/#bin.mongo

* Install MongoDB Database Tools (for JSON imports)

See:
https://docs.mongodb.com/database-tools/mongoimport
https://docs.mongodb.com/database-tools/installation/#install-tools

* Import prescription data into MongoDB

Our ultimate goal is to get the data into a [Kafka][1] data pipeline,
on [Confluent Cloud][2]. Confluent Cloud does not provide an out of the
box way of consuming JSON data (they instead require you to set up
additional infrastructure yourself). To avoid that overhead and to keep
things as simple as we can, we are using [MongoDB's Atlas cloud][3]
as our go-between for the JSON data and Kakfa. MongoDB is a good choice
of go-between as it:

1. [has an import tool which makes importing JSON easy][4]
2. [is a first-class connector on the Confluent Cloud platform][5],
which makes it straightforward to get the data from MongoDB into
Kafka.

[1]: https://kafka.apache.org/
[2]: https://www.confluent.io/confluent-cloud
[3]: https://www.mongodb.com/cloud/atlas
[4]: https://docs.mongodb.com/guides/server/import/
[5]: https://www.confluent.io/blog/confluent-cloud-offers-kafka-mongodb-connector-fully-managed-clusters/

* Update documentation

* Automate steps to add someone to SecretHub org

* Enable collaborators to add MongodB details

* Improve documentation for wrapper script

* Use MongoDB host rather than connection string

This which will make the upcoming integration with Kafka a little
smoother.

* Rename SecretHub template variable

env is a more versatile and descriptive name than user - e.g. the word
`user` does not describe `prod`, but the word `env` does.

* Give collaborators their own data url env var

This gives each collaborator their own full set of independent
environment variables, [in line with 12 Factor guidance][1].

It is important that each collaborator can experiment by changing the
data source url, independently of the other collaborators.

[1]: https://12factor.net/config

* Reorder sections in CONTRIBUTING file

* Import each record into MongoDB individually

This is important because to import the records properly into Kafka as
individual messages, we need them to be imported into MongoDB
individually.

We use [jq][1] to filter the source data to just the records array.

[1]: https://stedolan.github.io/jq/

* Improve script message which was ambiguous

* Improve database authentication documentation